Make Context generic
Adds a Generic parameter to Context (as well as Swd and Jtag) to carry the
Pins’ identity & state.
All pins are configured as SioOutput which are also InputPins.
`set_output_enable_override` is used to control the pin’s output control
at the pad-level. The Input circuitry always remains connected.

`Result::ok(self)` does not do any check and is misleading when use in the
form `.ok();` without using the resulting `Option<_>`. `let _ =` can
become heavy on the readability. This also crates a trait to ignore
OutputPin’s fallible result.